core: cli: cleanup main into subparts
